Add Podcast Owner Custom Definitions per #171
Some users were saying their podcasts were getting rejected since
`podcast.IOwner` wasn't being set.  This PR adds that functionality
to issue #171.

* Did not disturb existing custom Author string, as users may
already be using that (wasn't sure on the backwards compatibility
with this repo).
* "Name" needed a prefix, since there was already Title and
Description.  Picked `Owner` for now.
* Added to existing Unit Test to ensure it gets set when custom
config is used.

NOTE: I do not have a way to test this as I do not use this repo.
This is just a drive-by PR to help the authors.